+++ src/ShootyThing.java
@@ -2,7 +2,7 @@
 import java.awt.Graphics;
 
 	public class ShootyThing  extends GameObject{
-	int angle;
+	double angle;
 	int speed;
 		public ShootyThing(int x, int y, int width, int height) {
 			super(x, y, width, height);
@@ -13,13 +13,13 @@ import java.awt.Graphics;
 
 	 void draw(Graphics g) {
 		 g.setColor(Color.BLUE);
-	        g.fillRect(x, y, width, height);
+	        g.fillRect((int)x, (int)y, width, height);
 	     
 	 }
 	 void turn(double deg) {
 		 this.angle += deg;
 	 }
-	 int getAngle () {
+	 double getAngle () {
 		return angle;
 		 
 	 }
